requirements-gemini.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Gemini AI
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP
requests>=2.31.0

# Frontend
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0
